# feature/speed up release metrics calc